source: rename xtractor to source
